+++ src/driver/mysql/MysqlQueryRunner.ts
@@ -113,9 +113,6 @@ export class MysqlQueryRunner implements QueryRunner {
         } catch (error) {
             await this.rollbackTransaction();
             throw error;
-
-        } finally {
-            await this.release();
         }
     }