+++ src/driver/postgres/PostgresQueryRunner.ts
@@ -127,9 +127,6 @@ export class PostgresQueryRunner implements QueryRunner {
         } catch (error) {
             await this.rollbackTransaction();
             throw error;
-
-        } finally {
-            await this.release();
         }
     }