+++ src/driver/sqlserver/SqlServerDriver.ts
@@ -149,7 +149,7 @@ export class SqlServerDriver implements Driver {
         // pooling is enabled either when its set explicitly to true,
         // either when its not defined at all (e.g. enabled by default)
         return new Promise<void>((ok, fail) => {
-            const connection = new this.mssql.Connection(options).connect((err: any) => {
+            const connection = new this.mssql.ConnectionPool(options).connect((err: any) => {
                 if (err) return fail(err);
                 this.connectionPool = connection;
                 ok();